F12 add cost

+++ db/migrate/20171204140937_create_workflows.rb
@@ -2,7 +2,7 @@ class CreateWorkflows < ActiveRecord::Migration[5.1]
   def change
     create_table :workflows do |t|
       t.text :description
-      t.integer :hours
+      t.float :hours
 
       t.timestamps
     end
